Add `NumericAssertionsBase` to explicitly define nullability for Subject
