+++ Src/FluentAssertions/Numeric/NullableNumericAssertions.cs
@@ -7,6 +7,9 @@ using FluentAssertions.Execution;
 
 namespace FluentAssertions.Numeric;
 
+/// <summary>
+/// Contains a number of methods to assert that a nullable <see cref="IComparable{T}"/> is in the expected state.
+/// </summary>
 [DebuggerNonUserCode]
 public class NullableNumericAssertions<T> : NullableNumericAssertions<T, NullableNumericAssertions<T>>
     where T : struct, IComparable<T>
@@ -17,18 +20,24 @@ public class NullableNumericAssertions<T> : NullableNumericAssertions<T, Nullabl
     }
 }
 
+/// <summary>
+/// Contains a number of methods to assert that a nullable <see cref="IComparable{T}"/> is in the expected state.
+/// </summary>
 [DebuggerNonUserCode]
-public class NullableNumericAssertions<T, TAssertions> : NumericAssertions<T, TAssertions>
+public class NullableNumericAssertions<T, TAssertions> : NumericAssertionsBase<T, T?, TAssertions>
     where T : struct, IComparable<T>
     where TAssertions : NullableNumericAssertions<T, TAssertions>
 {
     private readonly AssertionChain assertionChain;
 
     public NullableNumericAssertions(T? value, AssertionChain assertionChain)
-        : base(value, assertionChain)
+        : base(assertionChain)
     {
+        Subject = value;
         this.assertionChain = assertionChain;
     }
+
+    public override T? Subject { get; }
 
     /// <summary>
     /// Asserts that a nullable numeric value is not <see langword="null"/>.
