+++ Src/FluentAssertions/NumericAssertionsExtensions.cs
@@ -34,7 +34,7 @@ public static class NumericAssertionsExtensions
         sbyte nearbyValue, byte delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        sbyte actualValue = parent.Subject.Value;
+        sbyte actualValue = parent.Subject;
         sbyte minValue = (sbyte)(nearbyValue - delta);
 
         if (minValue > nearbyValue)
@@ -77,7 +77,7 @@ public static class NumericAssertionsExtensions
         byte nearbyValue, byte delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        byte actualValue = parent.Subject.Value;
+        byte actualValue = parent.Subject;
         byte minValue = (byte)(nearbyValue - delta);
 
         if (minValue > nearbyValue)
@@ -120,7 +120,7 @@ public static class NumericAssertionsExtensions
         short nearbyValue, ushort delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        short actualValue = parent.Subject.Value;
+        short actualValue = parent.Subject;
         short minValue = (short)(nearbyValue - delta);
 
         if (minValue > nearbyValue)
@@ -164,7 +164,7 @@ public static class NumericAssertionsExtensions
         ushort nearbyValue, ushort delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        ushort actualValue = parent.Subject.Value;
+        ushort actualValue = parent.Subject;
         ushort minValue = (ushort)(nearbyValue - delta);
 
         if (minValue > nearbyValue)
@@ -208,7 +208,7 @@ public static class NumericAssertionsExtensions
         int nearbyValue, uint delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        int actualValue = parent.Subject.Value;
+        int actualValue = parent.Subject;
         int minValue = (int)(nearbyValue - delta);
 
         if (minValue > nearbyValue)
@@ -253,7 +253,7 @@ public static class NumericAssertionsExtensions
         uint nearbyValue, uint delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        uint actualValue = parent.Subject.Value;
+        uint actualValue = parent.Subject;
         uint minValue = nearbyValue - delta;
 
         if (minValue > nearbyValue)
@@ -298,7 +298,7 @@ public static class NumericAssertionsExtensions
         long nearbyValue, ulong delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        long actualValue = parent.Subject.Value;
+        long actualValue = parent.Subject;
         long minValue = GetMinValue(nearbyValue, delta);
         long maxValue = GetMaxValue(nearbyValue, delta);
 
@@ -331,7 +331,7 @@ public static class NumericAssertionsExtensions
         ulong nearbyValue, ulong delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        ulong actualValue = parent.Subject.Value;
+        ulong actualValue = parent.Subject;
         ulong minValue = nearbyValue - delta;
 
         if (minValue > nearbyValue)
@@ -390,7 +390,7 @@ public static class NumericAssertionsExtensions
         sbyte distantValue, byte delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        sbyte actualValue = parent.Subject.Value;
+        sbyte actualValue = parent.Subject;
         sbyte minValue = (sbyte)(distantValue - delta);
 
         if (minValue > distantValue)
@@ -434,7 +434,7 @@ public static class NumericAssertionsExtensions
         byte distantValue, byte delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        byte actualValue = parent.Subject.Value;
+        byte actualValue = parent.Subject;
         byte minValue = (byte)(distantValue - delta);
 
         if (minValue > distantValue)
@@ -479,7 +479,7 @@ public static class NumericAssertionsExtensions
         short distantValue, ushort delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        short actualValue = parent.Subject.Value;
+        short actualValue = parent.Subject;
         short minValue = (short)(distantValue - delta);
 
         if (minValue > distantValue)
@@ -524,7 +524,7 @@ public static class NumericAssertionsExtensions
         ushort distantValue, ushort delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        ushort actualValue = parent.Subject.Value;
+        ushort actualValue = parent.Subject;
         ushort minValue = (ushort)(distantValue - delta);
 
         if (minValue > distantValue)
@@ -568,7 +568,7 @@ public static class NumericAssertionsExtensions
         int distantValue, uint delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        int actualValue = parent.Subject.Value;
+        int actualValue = parent.Subject;
         int minValue = (int)(distantValue - delta);
 
         if (minValue > distantValue)
@@ -613,7 +613,7 @@ public static class NumericAssertionsExtensions
         uint distantValue, uint delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        uint actualValue = parent.Subject.Value;
+        uint actualValue = parent.Subject;
         uint minValue = distantValue - delta;
 
         if (minValue > distantValue)
@@ -657,7 +657,7 @@ public static class NumericAssertionsExtensions
         long distantValue, ulong delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        long actualValue = parent.Subject.Value;
+        long actualValue = parent.Subject;
         long minValue = GetMinValue(distantValue, delta);
         long maxValue = GetMaxValue(distantValue, delta);
 
@@ -690,7 +690,7 @@ public static class NumericAssertionsExtensions
         ulong distantValue, ulong delta,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        ulong actualValue = parent.Subject.Value;
+        ulong actualValue = parent.Subject;
         ulong minValue = distantValue - delta;
 
         if (minValue > distantValue)
@@ -851,17 +851,17 @@ public static class NumericAssertionsExtensions
 
         if (float.IsPositiveInfinity(expectedValue))
         {
-            FailIfDifferenceOutsidePrecision(float.IsPositiveInfinity(parent.Subject.Value), parent, expectedValue, precision,
+            FailIfDifferenceOutsidePrecision(float.IsPositiveInfinity(parent.Subject), parent, expectedValue, precision,
                 float.NaN, because, becauseArgs);
         }
         else if (float.IsNegativeInfinity(expectedValue))
         {
-            FailIfDifferenceOutsidePrecision(float.IsNegativeInfinity(parent.Subject.Value), parent, expectedValue, precision,
+            FailIfDifferenceOutsidePrecision(float.IsNegativeInfinity(parent.Subject), parent, expectedValue, precision,
                 float.NaN, because, becauseArgs);
         }
         else
         {
-            float actualDifference = Math.Abs(expectedValue - parent.Subject.Value);
+            float actualDifference = Math.Abs(expectedValue - parent.Subject);
 
             FailIfDifferenceOutsidePrecision(actualDifference <= precision, parent, expectedValue, precision, actualDifference,
                 because, becauseArgs);
@@ -989,17 +989,17 @@ public static class NumericAssertionsExtensions
 
         if (double.IsPositiveInfinity(expectedValue))
         {
-            FailIfDifferenceOutsidePrecision(double.IsPositiveInfinity(parent.Subject.Value), parent, expectedValue, precision,
+            FailIfDifferenceOutsidePrecision(double.IsPositiveInfinity(parent.Subject), parent, expectedValue, precision,
                 double.NaN, because, becauseArgs);
         }
         else if (double.IsNegativeInfinity(expectedValue))
         {
-            FailIfDifferenceOutsidePrecision(double.IsNegativeInfinity(parent.Subject.Value), parent, expectedValue, precision,
+            FailIfDifferenceOutsidePrecision(double.IsNegativeInfinity(parent.Subject), parent, expectedValue, precision,
                 double.NaN, because, becauseArgs);
         }
         else
         {
-            double actualDifference = Math.Abs(expectedValue - parent.Subject.Value);
+            double actualDifference = Math.Abs(expectedValue - parent.Subject);
 
             FailIfDifferenceOutsidePrecision(actualDifference <= precision, parent, expectedValue, precision, actualDifference,
                 because, becauseArgs);
@@ -1122,7 +1122,7 @@ public static class NumericAssertionsExtensions
     {
         Guard.ThrowIfArgumentIsNegative(precision);
 
-        decimal actualDifference = Math.Abs(expectedValue - parent.Subject.Value);
+        decimal actualDifference = Math.Abs(expectedValue - parent.Subject);
 
         FailIfDifferenceOutsidePrecision(actualDifference <= precision, parent, expectedValue, precision, actualDifference,
             because, becauseArgs);
@@ -1260,17 +1260,17 @@ public static class NumericAssertionsExtensions
 
         if (float.IsPositiveInfinity(unexpectedValue))
         {
-            FailIfDifferenceWithinPrecision(parent, !float.IsPositiveInfinity(parent.Subject.Value), unexpectedValue, precision,
+            FailIfDifferenceWithinPrecision(parent, !float.IsPositiveInfinity(parent.Subject), unexpectedValue, precision,
                 float.NaN, because, becauseArgs);
         }
         else if (float.IsNegativeInfinity(unexpectedValue))
         {
-            FailIfDifferenceWithinPrecision(parent, !float.IsNegativeInfinity(parent.Subject.Value), unexpectedValue, precision,
+            FailIfDifferenceWithinPrecision(parent, !float.IsNegativeInfinity(parent.Subject), unexpectedValue, precision,
                 float.NaN, because, becauseArgs);
         }
         else
         {
-            float actualDifference = Math.Abs(unexpectedValue - parent.Subject.Value);
+            float actualDifference = Math.Abs(unexpectedValue - parent.Subject);
 
             FailIfDifferenceWithinPrecision(parent, actualDifference > precision, unexpectedValue, precision, actualDifference,
                 because, becauseArgs);
@@ -1392,17 +1392,17 @@ public static class NumericAssertionsExtensions
 
         if (double.IsPositiveInfinity(unexpectedValue))
         {
-            FailIfDifferenceWithinPrecision(parent, !double.IsPositiveInfinity(parent.Subject.Value), unexpectedValue, precision,
+            FailIfDifferenceWithinPrecision(parent, !double.IsPositiveInfinity(parent.Subject), unexpectedValue, precision,
                 double.NaN, because, becauseArgs);
         }
         else if (double.IsNegativeInfinity(unexpectedValue))
         {
-            FailIfDifferenceWithinPrecision(parent, !double.IsNegativeInfinity(parent.Subject.Value), unexpectedValue, precision,
+            FailIfDifferenceWithinPrecision(parent, !double.IsNegativeInfinity(parent.Subject), unexpectedValue, precision,
                 double.NaN, because, becauseArgs);
         }
         else
         {
-            double actualDifference = Math.Abs(unexpectedValue - parent.Subject.Value);
+            double actualDifference = Math.Abs(unexpectedValue - parent.Subject);
 
             FailIfDifferenceWithinPrecision(parent, actualDifference > precision, unexpectedValue, precision, actualDifference,
                 because, becauseArgs);
@@ -1517,7 +1517,7 @@ public static class NumericAssertionsExtensions
     {
         Guard.ThrowIfArgumentIsNegative(precision);
 
-        decimal actualDifference = Math.Abs(unexpectedValue - parent.Subject.Value);
+        decimal actualDifference = Math.Abs(unexpectedValue - parent.Subject);
 
         FailIfDifferenceWithinPrecision(parent, actualDifference > precision, unexpectedValue, precision, actualDifference,
             because, becauseArgs);
@@ -1556,7 +1556,7 @@ public static class NumericAssertionsExtensions
         [StringSyntax("CompositeFormat")] string because = "",
         params object[] becauseArgs)
     {
-        float actualValue = parent.Subject.Value;
+        float actualValue = parent.Subject;
 
         parent.CurrentAssertionChain
             .ForCondition(float.IsNaN(actualValue))
@@ -1580,7 +1580,7 @@ public static class NumericAssertionsExtensions
         [StringSyntax("CompositeFormat")] string because = "",
         params object[] becauseArgs)
     {
-        double actualValue = parent.Subject.Value;
+        double actualValue = parent.Subject;
 
         parent.CurrentAssertionChain
             .ForCondition(double.IsNaN(actualValue))
@@ -1656,7 +1656,7 @@ public static class NumericAssertionsExtensions
         [StringSyntax("CompositeFormat")] string because = "",
         params object[] becauseArgs)
     {
-        float actualValue = parent.Subject.Value;
+        float actualValue = parent.Subject;
 
         parent.CurrentAssertionChain
             .ForCondition(!float.IsNaN(actualValue))
@@ -1680,7 +1680,7 @@ public static class NumericAssertionsExtensions
         [StringSyntax("CompositeFormat")] string because = "",
         params object[] becauseArgs)
     {
-        double actualValue = parent.Subject.Value;
+        double actualValue = parent.Subject;
 
         parent.CurrentAssertionChain
             .ForCondition(!double.IsNaN(actualValue))
